refactor(schema): move user table to schema-auth.ts for modularity
Isolated the User table definition into a separate file to improve maintainability and support custom auth without altering the main schema.ts
